\name{Bollen}
\alias{Bollen}
\docType{data}
\title{
Bollen's Data on Industrialization and Political Democracy
}
\description{
This data set includes four measures of democracy at two points in
time, 1960 and 1965, and three measures of industrialization in 1960,
for 75 developing countries.
}
\usage{Bollen}
\format{
  A data frame with 75 observations on the following 11 variables.
  \describe{
    \item{\code{y1}}{freedom of the press, 1960}
    \item{\code{y2}}{freedom of political opposition, 1960}
    \item{\code{y3}}{fairness of elections, 1960}
    \item{\code{y4}}{effectivness of elected legislature, 1960}
    \item{\code{y5}}{freedom of the press, 1965}
    \item{\code{y6}}{freedom of political opposition, 1965}
    \item{\code{y7}}{fairness of elections, 1965}
    \item{\code{y8}}{effectivness of elected legislature, 1965}
    \item{\code{x1}}{GNP per capita, 1960}
    \item{\code{x2}}{energy consumption per capita, 1960}
    \item{\code{x3}}{percentage of labor force in industry, 1960}
  }
}

\details{
Variables \code{y1} through \code{y4} are intended to be indicators of the latent variable \emph{political democracy in 1960};
\code{y5} through \code{y8} indicators of \emph{political democracy in 1965}; and
\code{x1} through \code{x3} indicators of \emph{industrialization in 1960}.
}

\source{
personal communication from Ken Bollen.
}
\references{
Bollen, K. A. (1989) \emph{Structural Equations With Latent Variables}. Wiley. 

}

\keyword{datasets}
