\name{Kmenta}
\alias{Kmenta}
\docType{data}
\title{Partly Artificial Data on the U. S. Economy}
\usage{Kmenta}
\description{
These are partly contrived data from Kmenta (1986), constructed
to illustrate estimation of a simultaneous-equation model.

The \code{Kmenta} data frame has 20 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Q}{
food consumption per capita.
}
\item{P}{
ratio of food prices to general consumer prices.
}
\item{D}{
disposable income in constant dollars.
}
\item{F}{
ratio of preceding year's prices received by farmers
to general consumer prices.
}
\item{A}{
time in years.
}
}
}

\details{
 The exogenous variables \code{D}, \code{F}, and \code{A} are based on
 real data; the endogenous variables \code{P} and \code{Q} were generated
 by simulation.
}
\source{
 Kmenta, J. (1986)
 \emph{Elements of Econometrics, Second Edition}, Macmillan.
}

\keyword{datasets}
